+++ tests/unit/utils/test_dict_utils.py
@@ -1,6 +1,10 @@
 import pytest
 
-from griptape.utils import dict_merge, remove_key_in_dict_recursively, remove_null_values_in_dict_recursively
+from griptape.utils import (
+    dict_merge,
+    remove_key_in_dict_recursively,
+    remove_null_values_in_dict_recursively,
+)
 
 
 class TestDictUtils: